refactor(extension): simplify file structure
