--- src/lib/element/aside.module.ts
@@ -1,8 +1,0 @@
-import { NgModule } from '@angular/core'
-import { AsideElement } from './aside'
-
-@NgModule({
-  declarations: [ AsideElement ],
-  exports: [ AsideElement ],
-})
-export class AsideElementModule { }
