--- src/lib/element/footer.module.ts
@@ -1,8 +1,0 @@
-import { NgModule } from '@angular/core'
-import { FooterElement } from './footer'
-
-@NgModule({
-  declarations: [ FooterElement ],
-  exports: [ FooterElement ],
-})
-export class FooterElementModule { }
