--- src/lib/element/header.module.ts
@@ -1,8 +1,0 @@
-import { NgModule } from '@angular/core'
-import { HeaderElement } from './header'
-
-@NgModule({
-  declarations: [ HeaderElement ],
-  exports: [ HeaderElement ],
-})
-export class HeaderElementModule { }
