--- src/lib/element/li.module.ts
@@ -1,8 +1,0 @@
-import { NgModule } from '@angular/core'
-import { LiElement } from './li'
-
-@NgModule({
-  declarations: [ LiElement ],
-  exports: [ LiElement ],
-})
-export class LiElementModule { }
