--- src/lib/element/main.module.ts
@@ -1,8 +1,0 @@
-import { NgModule } from '@angular/core'
-import { MainElement } from './main'
-
-@NgModule({
-  declarations: [ MainElement ],
-  exports: [ MainElement ],
-})
-export class MainElementModule { }
