--- src/lib/element/main.ts
@@ -1,9 +1,0 @@
-import { Directive } from '@angular/core'
-import { Element } from './element'
-
-@Directive({
-  selector: 'main:not([antNoEffect])',
-})
-export class MainElement extends Element {
-  tag = 'main'
-}
