+++ src/lib/layout/layout.module.ts
@@ -1,8 +1,5 @@
 import { NgModule } from '@angular/core'
-import { AsideElementModule } from '../element/aside.module'
-import { FooterElementModule } from '../element/footer.module'
-import { HeaderElementModule } from '../element/header.module'
-import { MainElementModule } from '../element/main.module'
+import { AsideElementModule, FooterElementModule, HeaderElementModule, MainElementModule } from '../element/element.module'
 import { ExtensionModule } from '../extension/extension.module'
 import { LayoutContent } from './content'
 import { LayoutFooter } from './footer'
